requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
supabase==2.4.3
httpx==0.27.0
python-multipart==0.0.9
weasyprint==62.1
jinja2==3.1.4
pandas==2.2.2
openpyxl==3.1.2
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
